+++ src/main/java/com/minhcraft/beyondbetatweaks/interfaces/IFractionalProtectionEnchantment.java
@@ -1,0 +1,8 @@
+package com.minhcraft.beyondbetatweaks.interfaces;
+
+import net.minecraft.world.damagesource.DamageSource;
+
+public interface IFractionalProtectionEnchantment {
+
+    float beyond_beta_tweaks$getFloatDamageProtection(int level, DamageSource source);
+}